% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npem_start.R
\name{npem.start}
\alias{npem.start}
\title{Obtain approximate starting values for npem.em}
\usage{
npem.start(y, cells = 10^6, n = c(24, 24, 24, 22), n.plates = 1,
  n.groups = 4, n.sd = 2, cv = 0.33)
}
\arguments{
\item{y}{Vector of transformed scintillation counts, in lexicographical
order (plate by plate and group by group within a plate.)}

\item{cells}{Number of cells per well.  The \eqn{\lambda}{lambda}'s will be
rescaled to give response per \eqn{10^6} cells.  This may be either a single
number (if all wells have the same number of cells, or \eqn{10^6} if one
wishes the \eqn{\lambda}{lambda}'s to not be rescaled), a value for each
plate (vector of length \code{n.plates}, or a value for each well (a vector
of the same length as \code{y}).}

\item{n}{Vector giving the number of wells within each group.  This may have
length either n.groups (if all plates have the same number of wells per
group) or n.groups*n.plates.}

\item{n.plates}{The number of plates in the data.}

\item{n.groups}{The number of groups. (This is needed here but not
elsewhere, because usually I figure it out from \code{n.plates} and the
length of the argument \code{ests}.)}

\item{n.sd}{Number of SDs above the mean to use as a cutoff}

\item{cv}{Coefficient of variation (= SD/ave) used in randomizing the
starting point; use cv=0 to avoid randomization.}
}
\value{
\item{ests}{The parameter estimates to use as starting values for
the EM algorithm, as a vector of length n.groups + 3*n.plates, of the form
(\eqn{\lambda}{lambda}'s, (a, b, \eqn{\sigma}{sigma})'s), where
\eqn{\lambda}{lambda} is the average number of responding cells per
\eqn{10^6} cells for a group, and (a, b, \eqn{\sigma}{sigma}) are the
plate-specific parameters.}
}
\description{
Obtains crude starting values, needed for the EM algorithm.
}
\details{
[I should describe the algorithm in more detail here.]
}
\examples{
  data(p713)
  start.pl3 <- npem.start(p713$counts[[3]],n=p713$n)
  out.pl3 <- npem.em(p713$counts[[3]],start.pl3,n=p713$n,tol=1e-13)

}
\references{
Broman et al. (1996) Estimation of antigen-responsive T cell
frequencies in PBMC from human subjects.  J Immunol Meth 198:119-132
}
\seealso{
\code{\link{npem.em}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{utilities}
